+++ releasenotes.md
@@ -71,6 +71,7 @@
 * Improvements have been made to the fuzz effect of the player’s weapon when they have a partial invisibility power-up.
 * A spectre now appears again in *DOOM II’s* cast sequence.
 * A bug has been fixed whereby the player would move slightly slower when using a gamepad rather than the keyboard.
+* The `+console` action may now be bound to a control on a gamepad.
 * Improvements have been made when the player uses the `kill` CCMD to commit suicide.
 
 ![](https://github.com/bradharding/www.doomretro.com/raw/master/wiki/bigdivider.png)